A couple of fixes for FSW (#239)
* MaxWait = 0 when a port has zero effects (was set to FOREVER)

* Only group together federated connections with same source port

* Correctly assign the connection ID of FederatedOutputConnection.
This was hardcoded to 0 which means it was bound to fail with multiple
connections between two federates.

* lf_set casts value to correct type.
There was an issue where we used the default  type of the literal
passed to `lf_set` this caused problems with float ports and passing
a literal e.g. 4.2 which is a double on my system.